Refactor and add types
